+++ src/index.ts
@@ -1,1 +1,3 @@
-export * from './SourceEmbedded'
+export * from './BridgeGuest'
+export * from './errors'
+export * from './types'